# gym-backend-api/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Weather API",
    "description": "JSON REST API for viewing weather data"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/bookings/details?pageNumber=&pageSize=": {
      "get": {
        "summary": "Gets all bookings with details",
        "description": "Endpoint to get all bookings with details.",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/bookings/{userID}": {
      "get": {
        "summary": "Get all bookings by user id",
        "description": "",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/bookings/{userID}/classes/{classID}": {
      "get": {
        "summary": "Check if a class is booked by a user",
        "description": "",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/bookings/number/{id}": {
      "get": {
        "summary": "Get future bookings number by user id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/booking": {
      "post": {
        "summary": "Creates a booking",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Creates a booking",
          "content": {
            "application/json": {
              "schema": {
                "booking_user_id": "string",
                "booking_class_id": "string",
                "booking_datetime": "string"
              },
              "example": {
                "booking_user_id": "1",
                "booking_class_id": "1",
                "booking_datetime": "2023-04-04 12:00:00"
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Updates a booking by id",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Updates a booking by id",
          "content": {
            "application/json": {
              "schema": {
                "booking_id": "string",
                "booking_user_id": "string",
                "booking_class_id": "string",
                "booking_created_datetime": "string"
              },
              "example": {
                "booking_id": "1",
                "booking_user_id": "1",
                "booking_class_id": "1",
                "booking_created_datetime": "2023-04-04 12:00:00"
              }
            }
          }
        }
      }
    },
    "/bookings/{bookingID}": {
      "delete": {
        "summary": "Deletes a booking by id",
        "description": "Endpoint to delete a booking by id.",
        "parameters": [
          {
            "name": "bookingID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}